pyproject.toml
[tool.poetry]
name = "e-ride"
version = "0.1.0"
description = "A ride hailing application"
authors = ["Ayomide <ayomidet905@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.0"
djangorestframework = "^3.14.0"
django-jazzmin = "^2.6.0"
djangorestframework-simplejwt = "^5.3.1"
python-decouple = "^3.8"
pytest = "^7.4.4"
pytest-django = "^4.7.0"
cloudinary = "^1.38.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-django = "^4.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
